+++ app/src/androidTest/java/dev/davidfdev/stela/ui/AdvancedSectionTest.kt
@@ -29,6 +29,8 @@ class AdvancedSectionTest {
                     onTitleChange = {},
                     onDescriptionChange = {},
                     onEmojiChange = {},
+                    onPinAtChange = {},
+                    onUnpinAtChange = {},
                     onTogglePin = {},
                     onToggleArchive = {},
                     onShare = {},